Integrate supplied typefaces and repair Caster hit delivery
Landed: seven static Barlow Condensed/Source Sans 3/Sometype Mono faces, licenses and rebuilt role assets; matching HUD draw/measure roles. Inventory and combat captures inspected at 1080p/720p.
Landed: actual projectile channel blocks on enemy combat bodies; Rot aims with WeaponTrace. Double-precision zone clocks retain final tick. Actual world diagnostic asserts 12 ticks and destruction.
Landed: five conditional Added Ability Power rows with legal rolls/equip and live damage/DoT snapshot coverage. Fracture300 and Rot10 O2 tuning retain original costs.
Measured: depth1 ten-second Rifle2058/Fracture2064; depth50 137527.5/130569.2. Full initial Mana bank and stationary target only. Rot retained trash activation1.5s, veteran4s, 12/12 ticks. Whole-loadout diagnostic parity0.648/0.771 preserves existing pinned fixtures.
Validation: final build/census pass; suite668 declared/started,665passed,3known,0unexpected via PowerShell quarantine auditor. Native Rift24kills/2015XP/209Glass and retained return rewards. STATE unchanged under host-Python quarantine.
Found-not-built: sustained build parity, authored encounter composition, wider polish, Void/Rift/reactions and multiplayer validation. No assertions or pins weakened. Initial build/test failures repaired before landing.

+++ Source/RiorsEdge/Combat/BreakerZoneActor.cpp
@@ -126,7 +126,7 @@ void ABreakerZoneActor::RefreshDuration(float NewDuration)
     if (!HasAuthority()) return;
     // Refresh, never extend: VW4's rule is that a recast resets the clock, so
     // spamming Rot cannot bank duration.
-    RemainingDuration = FMath::Max(RemainingDuration, FMath::Max(0.0f, NewDuration));
+    RemainingDuration = FMath::Max(RemainingDuration, static_cast<double>(FMath::Max(0.0f, NewDuration)));
 }
 
 void ABreakerZoneActor::SetFollowActor(AActor* Follow)
@@ -156,7 +156,7 @@ float ABreakerZoneActor::OwnedOccupiedSeconds(AActor* Owner, FGameplayTag Tag, f
         Zone->UpdateMembership();
         if (!Zone->Occupants.IsEmpty())
             Seconds = FMath::Max(Seconds, Zone->bExpiryPaused ? FMath::Max(0.0f, DeltaSeconds)
-                : FMath::Clamp(Zone->RemainingDuration, 0.0f, FMath::Max(0.0f, DeltaSeconds)));
+                : static_cast<float>(FMath::Clamp(Zone->RemainingDuration, 0.0, static_cast<double>(FMath::Max(0.0f, DeltaSeconds)))));
     }
     return Seconds;
 }
@@ -184,8 +184,8 @@ void ABreakerZoneActor::AdvanceZone(float DeltaSeconds)
 
     UpdateMembership();
 
-    const float ActiveSeconds = bExpiryPaused ? FMath::Max(0.0f, DeltaSeconds)
-        : FMath::Min(FMath::Max(0.0f, DeltaSeconds), FMath::Max(0.0f, RemainingDuration));
+    const double ActiveSeconds = bExpiryPaused ? FMath::Max(0.0f, DeltaSeconds)
+        : FMath::Min(static_cast<double>(FMath::Max(0.0f, DeltaSeconds)), FMath::Max(0.0, RemainingDuration));
     const int32 Ticks = UBreakerZoneMath::ConsumeTicks(TimeUntilNextTick, ActiveSeconds, Spec.TickInterval, MaximumTicksPerAdvance);
     for (int32 Index = 0; Index < Ticks; ++Index)
     {
